BartGadget: added auto download and building of libbartmain
Added possibility of automatically downloading, configuring and
building libbartmain.a as part of the compilation process.
This might be useful for people who do not have BART installed
or who do not remember with which options they compiled BART in
the first place.

With CMake >= 3.7 make use of the FetchContent module, otherwise
use the ExternalProject module to emulate that functionality as
best as possible.

Signed-off-by: Nguyen Damien <damien.nguyen@unibas.ch>

git-svn-id: http://10.92.115.44/svn/gadgetron/bart_gadget@3416 7c8698e1-dba2-45af-afb3-966a70042397